pyproject.toml
[project]
name = "claudini"
version = "0.1.0"
description = "Claudini: Automated Research Discovers State-of-the-Art Adversarial Attack Algorithms"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "numpy>=2.0",
    "pyyaml>=6.0",
    "scipy>=1.10",
    "torch>=2.0",
    "tqdm>=4.60",
    "transformers>=4.40",
    "typer>=0.9",
    "datasets>=2.14",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["claudini"]

[tool.ruff]
line-length = 120
target-version = "py314"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[project.optional-dependencies]
quantize = [
    "bitsandbytes>=0.43",
]

[project.scripts]
claudini-submit = "claudini.pipeline.submit:app"

[dependency-groups]
dev = [
    "ruff>=0.15.7",
]
